Hooked Callbacks into DeployHook
